add initial day 4 solution

+++ src/day_03/main.cpp
@@ -30,7 +30,7 @@ void read_input(std::vector<std::string>& lines) {
 int check_slope(std::vector<std::string> data, int along, int down) {
   int tree_count = 0;
   // slope
-  int offset = 0;
+  size_t offset = 0;
   for (auto i = down; i < data.size(); i += down) {
     auto& line = data[i];
     offset += along;
